+++ app/src/main/java/ru/vodolatskii/movies/data/service/KPApiService.kt
@@ -4,7 +4,7 @@ import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.Headers
 import retrofit2.http.Query
-import ru.vodolatskii.movies.data.dto.ShortDocsResponseDto
+import ru.vodolatskii.movies.data.entity.dto.ShortDocsResponseDto
 
 interface KPsApiService {
     @Headers(
