+++ app/src/main/java/ru/vodolatskii/movies/presentation/utils/contentRV/ContentAdapter.kt
@@ -50,6 +50,12 @@ class ContentAdapter(
         asyncListDiffer.submitList(list)
     }
 
+    fun addData(Movies: List<Movie>) {
+        val list = Movies.toMutableList() + getData()
+
+        asyncListDiffer.submitList(list)
+    }
+
     fun getData(): List<Movie> {
         return asyncListDiffer.currentList
     }
